激活邮件方法转移到邮件发送服务
